+++ sesac7_flutter_recipe_app/lib/data/mapper/ingredients_mapper.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_recipe_app/data/dto/ingredients_dto.dart';
 import 'package:flutter_recipe_app/data/mapper/ingredient_mapper.dart';
-import 'package:flutter_recipe_app/data/model/ingredients.dart';
+import 'package:flutter_recipe_app/domain/model/ingredients.dart';
 
 extension IngredientsDtoToModel on IngredientsDto {
   Ingredients toModel() {
@@ -15,7 +15,7 @@ extension IngredientsModelToDto on Ingredients {
   IngredientsDto toDto() {
     return IngredientsDto(
       ingredient: ingredient.toDto(),
-      amount: amount.toInt(),
+      amount: amount,
     );
   }
 }
